some norminette

+++ libft/ft_split.c
@@ -14,8 +14,8 @@
 
 static int	count_words(const char *str, char c)
 {
-	int i;
-	int trigger;
+	int	i;
+	int	trigger;
 
 	i = 0;
 	trigger = 0;
@@ -46,11 +46,11 @@ static char	*word_dup(const char *str, int start, int finish)
 	return (word);
 }
 
-char		**ft_split(char const *s, char c)
+char	**ft_split(char const *s, char c)
 {
 	int	i;
-	size_t	j;
-	int		index;
+	int	j;
+	int	index;
 	char	**split;
 
 	if (!s || !(split = malloc((count_words(s, c) + 1) * sizeof(char *))))
